feat(app): implement 404 page and refactor routing  Added the package-lock.json to the .gitignore. Added styles to the app.css for the NotFound.jsx file to display a 404 page when a user clicks on a page that does not exist. Added the correct routing for the app in the app.jsx. Included in components the exported functions that return HTML for the about.jsx, contact.jsx, portfolio.jsx, and resume.jsx. Added the code that handles checking which page the user is on in the NavTabs.jsx. Added a UI folder to include the form.jsx which is imported into the contact.jsx to assist with checking form submission parameters, and handling input change. Added a utils folder to touch a helpers.js file that assists the contact form with checking if email is entered correctly. Added the main.jsx code using the createRoot method from ReactDOM to render the App component onto the root element, included code to handle the VDOM, the App functionality, and the styling for the portfolio.
feat(app): implement 404 page and refactor routing

Added the package-lock.json to the .gitignore.
Added styles to the app.css for the NotFound.jsx file to display a 404 page when a user clicks on a page that does not exist.
Added the correct routing for the app in the app.jsx.
Included in components the exported functions that return HTML for the about.jsx, contact.jsx, portfolio.jsx, and resume.jsx.
Added the code that handles checking which page the user is on in the NavTabs.jsx.
Added a UI folder to include the form.jsx which is imported into the contact.jsx to assist with checking form submission parameters, and handling input change.
Added a utils folder to touch a helpers.js file that assists the contact form with checking if email is entered correctly.
Added the main.jsx code using the createRoot method from ReactDOM to render the App component onto the root element, included code to handle the VDOM, the App functionality, and the styling for the portfolio.