refactor(ui): consolidate useForm into Form component and update routes  Removed the UI folder and moved the useForm.jsx code into the Form.jsx code. Included the original Form.jsx code in the Contact.jsx so that the contact page functions correctly. Updated routing in NotFound.jsx to route back to the about page when the button is clicked. Changed routing in main.jsx; however, the initial page load does not take you to the About section and renders the site with no path. This issue will be fixed.
refactor(ui): consolidate useForm into Form component and update routes

Removed the UI folder and moved the useForm.jsx code into the Form.jsx code. Included the original Form.jsx code in the Contact.jsx so that the contact page functions correctly. Updated routing in NotFound.jsx to route back to the about page when the button is clicked. Changed routing in main.jsx; however, the initial page load does not take you to the About section and renders the site with no path. This issue will be fixed.