feat(header): format files and add global and component-specific styles  Added a global style to `index.css` and applied Bootstrap styling changes to `app.css`.  Wrapped the HTML return in `NavTabs.jsx` with a `<header>` to include the name in the navigation tabs.  Styled `NavTabs.css` to apply specific styles to the header and background to align the navbar with the header.  Included a font in the public folder (nested in a font folder) to style the name displayed on the portfolio header.
feat(header): format files and add global and component-specific styles

Added a global style to `index.css` and applied Bootstrap styling changes to `app.css`.
Wrapped the HTML return in `NavTabs.jsx` with a `<header>` to include the name in the navigation tabs.
Styled `NavTabs.css` to apply specific styles to the header and background to align the navbar with the header.
Included a font in the public folder (nested in a font folder) to style the name displayed on the portfolio header.