style(About.css): update styling and layout  Added new styling to the About.css to improve layout and responsiveness.  fix(Resume): add resume download link and viewer  Added the resume to the download button and included an iframe to display the resume on the resume page. Currently, the display is not functioning correctly and will be fixed in future updates.
style(About.css): update styling and layout

Added new styling to the About.css to improve layout and responsiveness.

fix(Resume): add resume download link and viewer

Added the resume to the download button and included an iframe to display the resume on the resume page. Currently, the display is not functioning correctly and will be fixed in future updates.