feat(projects): add project links and images to projectsData.jsx  Added all of the repo, live, and image links to the projectsData.jsx UI. Included relevant comments on each of the files. Updated README.md to include a link to the deployed site on Netlify. Removed Axios from package.json as it is redundant with emailjs in this configuration.
feat(projects): add project links and images to projectsData.jsx

Added all of the repo, live, and image links to the projectsData.jsx UI.
Included relevant comments on each of the files. Updated README.md to include a link to the deployed site on Netlify.
Removed Axios from package.json as it is redundant with emailjs in this configuration.